Moved packagin around to accomodate running+tests
