+++ main.py
@@ -1,6 +1,6 @@
 import sys
 from PyQt5.QtWidgets import QApplication
-from jide import jide
+from src.jide import jide
 
 if __name__ == "__main__":
     app = QApplication(sys.argv)
